v22.0.3

+++ src/backend/null.rs
@@ -1,7 +1,5 @@
 use crate::Backend;
 
-use autd3_driver::defined::float;
-
 /// Backend with no plotting
 pub struct NullBackend {}
 
@@ -20,9 +18,9 @@ impl Backend for NullBackend {
     }
 
     fn plot_1d(
-        _observe_points: Vec<float>,
+        _observe_points: Vec<f64>,
         _acoustic_pressures: Vec<autd3_driver::defined::Complex>,
-        _resolution: float,
+        _resolution: f64,
         _x_label: &str,
         _config: Self::PlotConfig,
     ) -> Result<(), crate::error::VisualizerError> {
@@ -30,10 +28,10 @@ impl Backend for NullBackend {
     }
 
     fn plot_2d(
-        _observe_x: Vec<float>,
-        _observe_y: Vec<float>,
+        _observe_x: Vec<f64>,
+        _observe_y: Vec<f64>,
         _acoustic_pressures: Vec<autd3_driver::defined::Complex>,
-        _resolution: float,
+        _resolution: f64,
         _x_label: &str,
         _y_label: &str,
         _config: Self::PlotConfig,
@@ -42,7 +40,7 @@ impl Backend for NullBackend {
     }
 
     fn plot_modulation(
-        _modulation: Vec<float>,
+        _modulation: Vec<f64>,
         _config: Self::PlotConfig,
     ) -> Result<(), crate::error::VisualizerError> {
         Err(crate::error::VisualizerError::NotSupported)
@@ -51,7 +49,7 @@ impl Backend for NullBackend {
     fn plot_phase(
         _config: Self::PlotConfig,
         _geometry: &autd3_driver::geometry::Geometry,
-        _phases: Vec<float>,
+        _phases: Vec<f64>,
     ) -> Result<(), crate::error::VisualizerError> {
         Err(crate::error::VisualizerError::NotSupported)
     }
